WebSite/BackEnd/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
sqlalchemy[asyncio]==2.0.34
asyncpg==0.29.0
alembic==1.13.2
aiosqlite==0.20.0
passlib[argon2]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
argon2-cffi==23.1.0
openai==1.40.6
httpx==0.27.0

